+++ MiniApp/Classes/core/Environment.swift
@@ -13,6 +13,7 @@ internal class Environment {
         case endpoint = "RMAAPIEndpoint"
         case isPreviewMode = "RMAIsPreviewMode"
         case hostAppUserAgentInfo = "RMAHostAppUserAgentInfo"
+        case requireMiniAppSignatureVerification = "RMARequireMiniAppSignatureVerification"
     }
 
     let bundle: EnvironmentProtocol
@@ -22,6 +23,7 @@ internal class Environment {
     var customAppVersion: String?
     var customSubscriptionKey: String?
     var customIsPreviewMode: Bool?
+    var customSignatureVerification: Bool?
 
     init(bundle: EnvironmentProtocol = Bundle.main) {
         self.bundle = bundle
@@ -29,19 +31,20 @@ internal class Environment {
 
     convenience init(with config: MiniAppSdkConfig, bundle: EnvironmentProtocol = Bundle.main) {
         self.init(bundle: bundle)
-        self.customUrl = config.baseUrl
-        self.customProjectId = config.rasProjectId
-        self.customSubscriptionKey = config.subscriptionKey
-        self.customAppVersion = config.hostAppVersion
-        self.customIsPreviewMode = config.isPreviewMode
+        customUrl = config.baseUrl
+        customProjectId = config.rasProjectId
+        customSubscriptionKey = config.subscriptionKey
+        customAppVersion = config.hostAppVersion
+        customIsPreviewMode = config.isPreviewMode
+        customSignatureVerification = config.requireMiniAppSignatureVerification
     }
 
     var projectId: String {
-        return value(for: customProjectId, fallback: .projectId)
+        value(for: customProjectId, fallback: .projectId)
     }
 
     var appVersion: String {
-        return value(for: customAppVersion, fallback: .version)
+        value(for: customAppVersion, fallback: .version)
     }
 
     var sdkVersion: MiniAppVersion? {
@@ -49,15 +52,19 @@ internal class Environment {
     }
 
     var subscriptionKey: String {
-        return value(for: customSubscriptionKey, fallback: .subscriptionKey)
+        value(for: customSubscriptionKey, fallback: .subscriptionKey)
     }
 
     var isPreviewMode: Bool {
-        return bool(for: customIsPreviewMode, fallback: .isPreviewMode)
+        bool(for: customIsPreviewMode, fallback: .isPreviewMode)
+    }
+
+    var requireMiniAppSignatureVerification: Bool {
+        bool(for: customSignatureVerification, fallback: .requireMiniAppSignatureVerification)
     }
 
     var hostAppUserAgentInfo: String {
-        return bundle.value(for: Key.hostAppUserAgentInfo.rawValue) ?? bundle.valueNotFound
+        bundle.value(for: Key.hostAppUserAgentInfo.rawValue) ?? bundle.valueNotFound
     }
 
     var baseUrl: URL? {
@@ -70,14 +77,14 @@ internal class Environment {
     }
 
     func value(for field: String?, fallback key: Key) -> String {
-        return field ?? bundle.value(for: key.rawValue) ?? bundle.valueNotFound
+        field ?? bundle.value(for: key.rawValue) ?? bundle.valueNotFound
     }
 
     func value(for field: String?, fallback key: Key, fallbackParam: String) -> String {
-        return field ?? bundle.value(for: key.rawValue) ?? fallbackParam
+        field ?? bundle.value(for: key.rawValue) ?? fallbackParam
     }
 
     func bool(for field: Bool?, fallback key: Key) -> Bool {
-        return field ?? bundle.bool(for: key.rawValue) ?? false
+        field ?? bundle.bool(for: key.rawValue) ?? false
     }
 }