+++ MiniApp/Classes/core/Extensions/URL+MiniApp.swift
@@ -9,4 +9,14 @@ extension URL {
         }
         return scheme?.starts(with: Constants.miniAppSchemePrefix) ?? false
     }
+
+    /// Replaces groups of non-alphanumeric characters in URL with '.'
+    /// ex. https://endpoint.com/v2/keys/ -> https.endpoint.com.v2.keys
+    var identifier: String {
+        absoluteString
+                .components(separatedBy: .alphanumerics.inverted)
+                .filter({ !$0.isEmpty })
+                .joined(separator: ".")
+                .trimmingCharacters(in: CharacterSet(charactersIn: "."))
+    }
 }