+++ MiniApp/Classes/core/Extensions/URLRequest+MiniApp.swift
@@ -1,9 +1,21 @@
 extension URLRequest {
-    static func createURLRequest(url: URL, environment: Environment) -> URLRequest {
+    static func createURLRequest(url: URL, subscriptionKey: String = "") -> URLRequest {
         var urlRequest = URLRequest(url: url)
-        if !environment.subscriptionKey.isEmpty {
-            urlRequest.setValue("ras-" + environment.subscriptionKey, forHTTPHeaderField: "apiKey")
+        if !subscriptionKey.isEmpty {
+            urlRequest.setSubscriptionKey(subscriptionKey)
         }
         return urlRequest
+    }
+
+    static func createURLRequest(url: URL, environment: Environment) -> URLRequest {
+        createURLRequest(url: url, subscriptionKey: environment.subscriptionKey)
+    }
+
+    mutating func setSubscriptionKey(_ key: String) {
+        var keyValue = key
+        if !key.hasPrefix("ras-") {
+            keyValue = "ras-\(key)"
+        }
+        setValue(keyValue, forHTTPHeaderField: "apiKey")
     }
 }