+++ MiniApp/Classes/core/MiniApp.swift
@@ -1,7 +1,7 @@
 /// Mini App Public API methods
 public class MiniApp: NSObject {
     private static let shared = MiniApp()
-    private let realMiniApp = RealMiniApp()
+    let realMiniApp = RealMiniApp()
     public static var MAOrientationLock: UIInterfaceOrientationMask = []
 
     private override init() {