+++ MiniApp/Classes/signature/Extensions/Data+MiniApp.swift
@@ -1,0 +1,15 @@
+import CommonCrypto
+
+extension Data {
+    func sha256() -> Data {
+        var hash = [UInt8](repeating: 0, count: Int(CC_SHA256_DIGEST_LENGTH))
+        self.withUnsafeBytes {
+            _ = CC_SHA256($0.baseAddress, CC_LONG(count), &hash)
+        }
+        return Data(hash)
+    }
+
+    func sha256String() -> String {
+        sha256().map { String(format: "%02hhx", $0) }.joined()
+    }
+}